// frontend/src/interfaces/NutritionalValue.ts
export interface INutritionalValue {
  id: number;
  recipeId: number;
  adjustedProtein: string;
  protein: string;
  carbohydrate: string;
  totalFat: string;
  totalDietaryFiber: string;
  totalSugars: string;
  energyKcal: string;
  energyKj: string;
  alanine: string;
  arginine: string;
  asparticAcid: string;
  cystine: string;
  glutamicAcid: string;
  glycine: string;
  histidine: string;
  hydroxyproline: string;
  isoleucine: string;
  leucine: string;
  lysine: string;
  methionine: string;
  phenylalanine: string;
  proline: string;
  serine: string;
  threonine: string;
  tryptophan: string;
  tyrosine: string;
  valine: string;
  calcium: string;
  copper: string;
  iron: string;
  magnesium: string;
  manganese: string;
  phosphorus: string;
  potassium: string;
  selenium: string;
  sodium: string;
  zinc: string;
  vitaminAIu: string;
  vitaminARae: string;
  vitaminB12: string;
  vitaminB12Added: string;
  vitaminB6: string;
  vitaminC: string;
  vitaminD: string;
  vitaminDIu: string;
  vitaminD2: string;
  vitaminD3: string;
  vitaminE: string;
  vitaminEAdded: string;
  vitaminK: string;
  niacin: string;
  pantothenicAcid: string;
  riboflavin: string;
  thiamin: string;
  folateDfe: string;
  folateFood: string;
  folateTotal: string;
  folicAcid: string;
  totalSaturatedFattyAcids: string;
  saturatedFattyAcids40: string;
  saturatedFattyAcids60: string;
  saturatedFattyAcids80: string;
  saturatedFattyAcids100: string;
  saturatedFattyAcids120: string;
  saturatedFattyAcids140: string;
  saturatedFattyAcids141: string;
  saturatedFattyAcids160: string;
  saturatedFattyAcids161: string;
  saturatedFattyAcids180: string;
  saturatedFattyAcids181: string;
  saturatedFattyAcids182: string;
  saturatedFattyAcids183: string;
  saturatedFattyAcids184: string;
  saturatedFattyAcids200: string;
  saturatedFattyAcids201: string;
  saturatedFattyAcids204: string;
  saturatedFattyAcids205Epa: string;
  saturatedFattyAcids220: string;
  saturatedFattyAcids221: string;
  saturatedFattyAcids225Dpa: string;
  saturatedFattyAcids226Dha: string;
  totalMonounsaturatedFattyAcids: string;
  totalPolyunsaturatedFattyAcids: string;
  polyunsaturatedFattyAcids150: string;
  polyunsaturatedFattyAcids161c: string;
  polyunsaturatedFattyAcids161t: string;
  polyunsaturatedFattyAcids170: string;
  polyunsaturatedFattyAcids171: string;
  polyunsaturatedFattyAcids181c: string;
  polyunsaturatedFattyAcids181t: string;
  polyunsaturatedFattyAcids182Clas: string;
  polyunsaturatedFattyAcids182i: string;
  polyunsaturatedFattyAcids182n6cc: string;
  polyunsaturatedFattyAcids182tNotDefined: string;
  polyunsaturatedFattyAcids182tt: string;
  polyunsaturatedFattyAcids183n6ccc: string;
  polyunsaturatedFattyAcids202n6cc: string;
  polyunsaturatedFattyAcids203: string;
  polyunsaturatedFattyAcids221c: string;
  polyunsaturatedFattyAcids221t: string;
  polyunsaturatedFattyAcids240: string;
  polyunsaturatedFattyAcids241c: string;
  totalTransFattyAcids: string;
  totalTransMonoenoicFattyAcids: string;
  totalTransPolyenoicFattyAcids: string;
  transPolyenoicFattyAcids130: string;
  transPolyenoicFattyAcids151: string;
  transPolyenoicFattyAcids18111t: string;
  transPolyenoicFattyAcids183n3cccAla: string;
  transPolyenoicFattyAcids183i: string;
  transPolyenoicFattyAcids203n3: string;
  transPolyenoicFattyAcids203n6: string;
  transPolyenoicFattyAcids204n6: string;
  transPolyenoicFattyAcids215: string;
  transPolyenoicFattyAcids224: string;
  fructose: string;
  galactose: string;
  glucose: string;
  lactose: string;
  maltose: string;
  starch: string;
  sucrose: string;
  alphaCarotene: string;
  betaCarotene: string;
  betaCryptoxanthin: string;
  luteinZeaxanthin: string;
  lycopene: string;
  betaTocopherol: string;
  deltaTocopherol: string;
  gammaTocopherol: string;
  alphaTocotrienol: string;
  betaTocotrienol: string;
  deltaTocotrienol: string;
  gammaTocotrienol: string;
  betaSitosterol: string;
  campesterol: string;
  cholesterol: string;
  phytosterols: string;
  stigmasterol: string;
  ethylAlcohol: string;
  ash: string;
  betaine: string;
  caffeine: string;
  totalCholine: string;
  dihydrophylloquinone: string;
  fluoride: string;
  menaquinone4: string;
  retinol: string;
  theobromine: string;
  water: string;
}